requirements.txt
boto3
python-dotenv
pandas
sqlalchemy
pymysql
